# server/docs/swagger.yaml
definitions:
...
  request.CreateAuditLogRequest:
    properties:
      action:
        description: 操作动作
        type: string
      category:
        description: 事件分类
        type: string
      client_id_initiator:
        description: 发起方客户端ID
        type: string
      client_id_responder:
        description: 响应方客户端ID
        type: string
      details:
        description: 事件详情
      duration:
        description: 操作耗时(毫秒)
        type: integer
      error_message:
        description: 错误信息
        type: string
      event_type:
        description: 事件类型
        type: string
      level:
        description: 日志级别
        type: string
      request_id:
        description: 请求ID
        type: string
      resource:
        description: 操作资源
        type: string
      result:
        description: 操作结果
        type: string
      server_id:
        description: 服务器ID
        type: string
      session_id:
        description: 会话ID
        type: string
      source_ip:
        description: 源IP地址
        type: string
      user_agent:
        description: 用户代理
        type: string
      user_id:
        description: 用户ID
        type: string
    required:
    - event_type
    - level
    - result
    type: object
...
  response.AuditLogEntry:
    properties:
      action:
        description: 操作动作
        type: string
      category:
        description: 事件分类
        type: string
      client_id_initiator:
        description: 发起方客户端ID
        type: string
      client_id_responder:
        description: 响应方客户端ID
        type: string
      created_at:
        description: 创建时间
        type: string
      details:
        description: 事件详情JSON
        type: string
      duration:
        description: 操作耗时(毫秒)
        type: integer
      error_message:
        description: 错误信息
        type: string
      event_time:
        description: 事件发生时间
        type: string
      event_type:
        description: 事件类型
        type: string
      id:
        description: 日志ID
        type: integer
      level:
        description: 日志级别
        type: string
      request_id:
        description: 请求ID
        type: string
      resource:
        description: 操作资源
        type: string
      result:
        description: 操作结果
        type: string
      server_id:
        description: 服务器ID
        type: string
      session_id:
        description: 会话ID
        type: string
      source_ip:
        description: 源IP地址
        type: string
      user_agent:
        description: 用户代理
        type: string
      user_id:
        description: 用户ID
        type: string
    type: object
...
  response.ClientDetailsResponse:
    properties:
      client_id:
        description: 客户端ID
        type: string
      connected_at:
        description: 连接时间
        type: string
      connection_events:
        description: 连接事件历史
        items:
          $ref: '#/definitions/response.ConnectionEvent'
        type: array
      display_name:
        description: 显示名称
        type: string
      ip_address:
        description: IP地址
        type: string
      is_online:
        description: 在线状态
        type: boolean
      last_message_at:
        description: 最后消息时间
        type: string
      received_message_count:
        description: 已接收消息数
        type: integer
      role:
        description: 角色：provider/receiver/none
        type: string
      sent_message_count:
        description: 已发送消息数
        type: integer
      session_id:
        description: 当前会话ID
        type: string
      user_agent:
        description: 用户代理
        type: string
      user_id:
        description: 用户ID
        type: string
    type: object
  response.ClientInfo:
    properties:
      client_id:
        description: 客户端ID
        type: string
      connected_at:
        description: 连接时间
        type: string
      display_name:
        description: 显示名称
        type: string
      ip_address:
        description: IP地址
        type: string
      is_online:
        description: 在线状态
        type: boolean
      role:
        description: 角色：provider/receiver/none
        type: string
      session_id:
        description: 当前会话ID
        type: string
      user_id:
        description: 用户ID
        type: string
    type: object
...
  response.ConnectionEvent:
    properties:
      client_id:
        description: 相关客户端ID
        type: string
      details:
        description: 事件详情
        type: string
      event:
        description: 事件类型
        type: string
      timestamp:
        description: 事件时间
        type: string
    type: object
...
  response.PaginatedAuditLogResponse:
    properties:
      list:
        description: 审计日志列表
        items:
          $ref: '#/definitions/response.AuditLogEntry'
        type: array
      page:
        description: 当前页码
        type: integer
      pageSize:
        description: 每页条数
        type: integer
      total:
        description: 总数
        type: integer
    type: object
...
  response.PaginatedClientListResponse:
    properties:
      list:
        description: 客户端列表
        items:
          $ref: '#/definitions/response.ClientInfo'
        type: array
      page:
        description: 当前页码
        type: integer
      pageSize:
        description: 每页数量
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.2
paths:
  /admin/nfc-relay/v1/audit-logs:
    get:
      consumes:
      - application/json
      description: 获取审计日志列表，支持分页和筛选
      parameters:
      - description: 事件分类
        in: query
        name: category
        type: string
      - description: 客户端ID
        in: query
        name: clientID
        type: string
      - description: 结束时间（ISO8601格式）
        in: query
        name: endTime
        type: string
      - description: 事件类型
        in: query
        name: eventType
        type: string
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      - description: 日志级别
        in: query
        name: level
        type: string
      - description: 页码
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 每页条数
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      - description: 操作结果
        in: query
        name: result
        type: string
      - description: 会话ID
        in: query
        name: sessionID
        type: string
      - description: 源IP
        in: query
        name: sourceIP
        type: string
      - description: 开始时间（ISO8601格式）
        in: query
        name: startTime
        type: string
      - description: 用户ID
        in: query
        name: userID
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PaginatedAuditLogResponse'
              type: object
      summary: 获取审计日志列表
      tags:
      - NFC中继审计日志
    post:
      consumes:
      - application/json
      description: 创建新的审计日志记录
      parameters:
      - description: 创建审计日志
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CreateAuditLogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/response.Response'
      summary: 创建审计日志
      tags:
      - NFC中继审计日志
...
  /admin/nfc-relay/v1/clients:
    get:
      consumes:
      - application/json
      description: 支持分页和筛选的客户端列表查询
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 客户端ID筛选
        in: query
        name: clientID
        type: string
      - description: 用户ID筛选
        in: query
        name: userID
        type: string
      - description: 角色筛选
        in: query
        name: role
        type: string
      - description: IP地址筛选
        in: query
        name: ipAddress
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PaginatedClientListResponse'
              type: object
      summary: 获取连接客户端列表
      tags:
      - NFC中继管理
  /admin/nfc-relay/v1/clients/{clientID}/details:
    get:
      consumes:
      - application/json
      description: 获取客户端的详细信息包括连接历史和统计
      parameters:
      - description: 客户端ID
        in: path
        name: clientID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.ClientDetailsResponse'
              type: object
      summary: 获取指定客户端的详细信息
      tags:
      - NFC中继管理
  /admin/nfc-relay/v1/clients/{clientID}/disconnect:
    post:
      consumes:
      - application/json
      description: 强制断开指定客户端的WebSocket连接
      parameters:
      - description: 客户端ID
        in: path
        name: clientID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 管理员强制断开指定客户端
      tags:
      - NFC中继管理
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...